+++ R/getMeta.R
@@ -39,6 +39,11 @@
 #' @param duplicate Some UK air quality sites are part of multiple networks, so
 #'   could appear more than once when `source` is a vector of two or more. The
 #'   default argument, `FALSE`, drops duplicate sites. `TRUE` will return them.
+#' @param year If a single year is selected, only sites that were open at some
+#'   point in that year are returned. If \code{all = TRUE} only sites that
+#'   measured a particular pollutant in that year are returned. Year can also be
+#'   of length 2 e.g. \code{year = c(2010, 2020)}, which will return only sites
+#'   that were open over the duration.
 #' @return A data frame with meta data.
 #' @author David Carslaw
 #' @family import functions
@@ -62,7 +67,7 @@
 #' meta <- importMeta(source = c("aurn", "aqe", "local"))
 #' }
 
-importMeta <- function(source = "aurn", all = FALSE, duplicate = FALSE) {
+importMeta <- function(source = "aurn", all = FALSE, duplicate = FALSE, year = NA) {
 
   ## meta data sources
   meta.source <-
@@ -91,7 +96,7 @@ importMeta <- function(source = "aurn", all = FALSE, duplicate = FALSE) {
         lmam = "https://uk-air.defra.gov.uk/openair/LMAM/R_data/LMAM_metadata.RData"
       )
 
-      meta <- clean_ricardo_meta(url, all = all)
+      meta <- clean_ricardo_meta(url, all = all, year = year)
     }
 
     # KCL
@@ -189,8 +194,9 @@ importMeta <- function(source = "aurn", all = FALSE, duplicate = FALSE) {
 #' Clean data from Ricardo (not KCL or Europe)
 #' @param url URL to use
 #' @param all inherited from parent function
+#' @param year inherited from parent function
 #' @noRd
-clean_ricardo_meta <- function(url, all) {
+clean_ricardo_meta <- function(url, all, year) {
   tmp <- tempfile()
 
   # load data
@@ -219,6 +225,23 @@ clean_ricardo_meta <- function(url, all) {
   if ("ratified_to" %in% names(meta)) {
     meta$ratified_to <-
       lubridate::ymd(meta$ratified_to, tz = "GMT", quiet = TRUE)
+  }
+  
+  # select year or period when sites were open
+  if (!anyNA(year)) {
+    
+    # format end_date - set "ongoing" to current date
+    meta$end_date[which(meta$end_date == "ongoing")] <- as.character(Sys.Date())
+    meta$end_year <- lubridate::year(as.Date(meta$end_date))
+    meta$start_year <- lubridate::year(meta$start_date)
+    
+    if (length(year) == 1) 
+      meta <- filter(meta, start_year <= year & end_year >= year)
+    
+    if (length(year) == 2) 
+      meta <- filter(meta, start_year <= year[1] & end_year >= year[2])
+    
+    
   }
 
   ## only extract one line per site to make it easier to use file